requirements.txt
groq
langchain
langchain-community
langchain-huggingface
langchain-groq
langgraph
sentence-transformers
torch
pinecone
langchain-pinecone
tavily-python
newspaper3k
beautifulsoup4
lxml
lxml-html-clean
requests
easyocr
pytesseract
Pillow
fastapi
uvicorn[standard]
python-multipart
streamlit
pydantic
pydantic-settings
python-dotenv
ragas
datasets
langsmith
pandas
numpy
tqdm
httpx
langchain-sambanova
langchain-google-genai
slowapi
limits[redis]
redis
jinja2
